+++ C# OOP Advanced/ExamPreps/Hell/Hell-Skeleton/Hell/Entities/Items/CommonItem.cs
@@ -1,6 +1,6 @@
 ﻿public class CommonItem : AbstracItem
 {
-    public CommonItem(string name, int strenghtBonus, int agilityBonus, int inteligenceBonus, int hitpointsBonus, int demageBonus)
+    public CommonItem(string name, long strenghtBonus, long agilityBonus, long inteligenceBonus, long hitpointsBonus, long demageBonus)
         : base(name, strenghtBonus, agilityBonus, inteligenceBonus, hitpointsBonus, demageBonus)
     {
     }
